moved equipment_types.py to game/ directory
