+++ game/components/equippable.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from equipment_types import EquipmentType
+from game.equipment_types import EquipmentType
 from game.components.base_component import BaseComponent
 import game.entity
